pyproject.toml
[project]
name = "toxic-comments-severity"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "bryan_goh_yc", email = "bryan_goh_yc@aiap.sg"},
]
dependencies = [
    "matplotlib>=3.8.3",
    "seaborn>=0.13.2",
    "scikit-learn>=1.4.1.post1",
    "pandas>=2.2.1",
    "jupyter>=1.0.0",
    "polars>=0.20.14",
    "gradio>=4.21.0",
    "transformers[torch]>=4.38.2",
    "evaluate>=0.4.1",
    "hydra-core>=1.3.2",
    "torch>=2.2.1",
    "torchvision>=0.17.1",
    "torchaudio>=2.2.1",
    "accelerate>=0.27.2",
    "huggingface-hub>=0.21.4",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
